Explicitly namespace-qualify all std names in header files
The windows executables fail to compile with gcc 11 mingw cross-compile
toolchain which defaults to '-std=c++17', due to conflicting definitions.

  In file included from /usr/x86_64-w64-mingw32/sys-root/mingw/include/wtypes.h:8,
                   from /usr/x86_64-w64-mingw32/sys-root/mingw/include/winscard.h:10,
                   from /usr/x86_64-w64-mingw32/sys-root/mingw/include/windows.h:97,
                   from /usr/x86_64-w64-mingw32/sys-root/mingw/include/rpc.h:16,
                   from guid.h:23,
                   from parttypes.h:16,
                   from gptpart.h:22,
                   from gptpart.cc:27:
  /usr/x86_64-w64-mingw32/sys-root/mingw/include/rpcndr.h:64:11:
  error:reference to 'byte' is ambiguous
     64 |   typedef byte cs_byte;
        |           ^~~~
  In file included from /usr/x86_64-w64-mingw32/sys-root/mingw/include/c++/bits/stl_algobase.h:61,
                   from /usr/x86_64-w64-mingw32/sys-root/mingw/include/c++/bits/char_traits.h:39,
                   from /usr/x86_64-w64-mingw32/sys-root/mingw/include/c++/ios:40,
                   from /usr/x86_64-w64-mingw32/sys-root/mingw/include/c++/ostream:38,
                   from /usr/x86_64-w64-mingw32/sys-root/mingw/include/c++/iostream:39,
                   from gptpart.cc:26:
  /usr/x86_64-w64-mingw32/sys-root/mingw/include/c++/bits/cpp_type_traits.h:404:30: note: candidates are: 'enum class std::byte'
    404 |   enum class byte : unsigned char;
        |                              ^~~~

The issue has been unnoticed so far all the way up to gcc 10, which
defaulted to '-std=c++14'. But it can still be reproduced with gcc 10
with an explicit '-std=c++17' flag.

"using namespace std" in header files is considered as a bad practice,
often known as global namespace pollution. Problems may occur when more
than one namespace has the same symbol name with signature, then it will
be ambiguous for the compiler to decide which one to use.

Let's explicitly namespace-qualify all std names in header files, and
move "using namespace std" to *.cc files.

Signed-off-by: Bin Meng <bmeng.cn@gmail.com>