makefile: Avoid absolute path in CXX for mingw* builds
At present CXX specifies absolute path for the MinGW GCC compiler.
But it only applies to MinGW cross compiler on the Linux host.
On a Windows host with MSYS2, MinGW GCC is normally installed at
/mingw{32,64}/bin which does not work with current makefile.

Remove the absolute path prefix, as on both Linux and Windows hosts
MinGW GCC compiler path is already in the PATH environment.

Signed-off-by: Bin Meng <bmeng.cn@gmail.com>