README: Update build instructions for Windows executables
Update the Windows documentation to mention:
- sgdisk{32,64}.exe is now provided
- MSYS2 Windows-hosted MinGW compiler is verified
- instructions to compile on Windows and cross-compile on Linux

Signed-off-by: Bin Meng <bmeng.cn@gmail.com>

Series-to: Rod Smith <rodsmith@rodsbooks.com>, gptfdisk-general@lists.sourceforge.net
Cover-letter:
Clean up makefiles and support building sgdisk executable for Windows
This series was originated from trying to build the sgdisk executable
for Windows host, that has not been formally supported as of today.

The series starts from several patches to do some clean-ups to existing
makefiles for various platforms (patch 1, 2, 3, 4, 6, 7).

Patch 5 is a build warning fix as the result of adding the '-O2' flag.

Patch 8 fixes several issues in the 'sgdisk' build rule for mingw*
builds, and patch 9 enables the sgdisk build by default.

Path 10 does further clean-up to simplify the build rules.

Path 11 and 12 are fixes when using newer version gcc (e.g.: gcc 11)
which is the one that ships on MSYS2.

Patch 13 switches the Windows executable build from static to shared.

Patch 14 allows MinGW compiler on both Linux and Windows to be used.

Patch 15 updates the Windows documentation with the latest information.
END